+++ tests/functional/auth/test_auth_client_flow.py
@@ -12,9 +12,9 @@ INVALID_GRANT_RESPONSE_BODY = '{"error":"invalid_grant"}'
 
 
 @pytest.fixture
-def client(no_retry_policy):
+def client(no_retry_transport):
     class CustomAuthClient(globus_sdk.AuthClient):
-        retry_policy = no_retry_policy
+        transport_class = no_retry_transport
 
     return CustomAuthClient(client_id=CLIENT_ID)
